+++ src/client/components/EndSession.jsx
@@ -4,7 +4,7 @@ import { useLocation } from 'react-router-dom';
 const EndSession = () => {
 	const location = useLocation();
 	useEffect(() => {
-		window.location.href = `/api/v1/app/login?redirect=${location.pathname}`;
+		window.location.href = `/api/v1/session/login?redirect=${location.pathname}`;
 	}, []);
 	return (<Fragment key="empty" />);
 };
