+++ src/client/components/SignUpForm/SignUpForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { logOut } from '../../utils/api/app/session';
+import { logOut } from '../../utils/api/session';
 import { createUser } from '../../utils/api/users';
 import './styles.css';
