+++ src/server/server.js
@@ -6,13 +6,13 @@ const morgan = require('morgan');
 const path = require('path');
 const sequelize = require('./models');
 const router = require('./router');
-const appRouter = require('./routers/app');
 const locationsRouter = require('./routers/locations');
 const lockersRouter = require('./routers/lockers');
 const usersRouter = require('./routers/users');
 const rentalsRouter = require('./routers/rentals');
 const paymentMethodsRouter = require('./routers/paymentMethods');
 const paymentsRouter = require('./routers/payments');
+const sessionRouter = require('./routers/session');
 const config = require('./config/server.json')[env];
 const { globalErrorHandler, NotFoundError } = require('./errors');
 
@@ -49,7 +49,7 @@ app.use(session({
 	},
 }));
 // Main API router.
-app.use('/api/v1/app', appRouter);
+app.use('/api/v1/session', sessionRouter);
 app.use('/api/v1/locations', locationsRouter);
 app.use('/api/v1/lockers', lockersRouter);
 app.use('/api/v1/users', usersRouter);
